Create receiptEmail
This file was from when I was just getting started in VisualBasic. It's been one of the more complex things I've done to generate an email in Outlook and a PDF from a Word Mail merge, open the PDF, and attach it to the email that's already addressed to the correct recipient from the correct inbox.